updating all metrics to float64 as per https://github.com/open-meteo/open-meteo/issues/410

+++ parsing.go
@@ -35,7 +35,7 @@ type Forecast struct {
 type CurrentWeather struct {
 	Temperature   float64
 	Time          ApiTime
-	WeatherCode   int
+	WeatherCode   float64
 	WindDirection float64
 	WindSpeed     float64
 }
